Finally finished explorer mode, jumped ahead to Epic mode without doing adventure mode, but will finish adventure mode over the weekend.

+++ Client.cs
@@ -4,7 +4,7 @@ namespace FirstBankOfSuncoast
     {
         public string UserName { get; set; }
         public int Pin { get; set; }
-        public double SavingsBalance { get; set; }
-        public double CheckingsBalance { get; set; }
+        public int SavingsBalance { get; set; }
+        public int CheckingsBalance { get; set; }
     }
 }
